FIX :: types 변경 - 두번 이상 쓰이는것들을 types안에 넣었습니다 - 한번만 쓰이는것은 컴포넌트 상단에 interface로 박았습니다
