+++ src/components/layout/Auth/Login/index.tsx
@@ -1,5 +1,5 @@
 import { ChangeEvent, useState } from 'react';
-import { LoginType } from 'types/auth/auth.type';
+import { LoginType } from 'types/auth.type';
 import Input from 'components/shared/Input';
 import Button from 'components/shared/Button';
 import { useNavigate } from 'react-router-dom';
