+++ src/features/auth/login.feature.ts
@@ -5,7 +5,7 @@ import { useMutation } from 'react-query';
 import { loginUser } from 'apis/auth.api';
 import { useSetRecoilState } from 'recoil';
 import { userData } from 'atoms/user.atom';
-import { LoginType } from 'types/auth/auth.type';
+import { LoginType } from 'types/auth.type';
 
 export interface LoginFeatureType {
   loginData: LoginType;
