+++ src/types/common/button.type.ts
@@ -1,8 +1,10 @@
 import { ButtonHTMLAttributes } from 'react';
 
+export type ButtonOptionType = 'UNFILLED' | 'FILLED' | 'SCARCE_FILLED';
+
 export interface ButtonPropsType
   extends ButtonHTMLAttributes<HTMLButtonElement> {
   padding?: string;
   width?: string;
-  option: 'UNFILLED' | 'FILLED' | 'SCARCE_FILLED';
+  option: ButtonOptionType;
 }
