+++ src/types/common/progress.type.ts
@@ -1,0 +1,1 @@
+export type ProgressBarOption = 'LIST' | 'DETAIL';
